DST-381: update cya bugfixes (#110)
1) Non-inputted fields showing up on check your answers

Business you work for shows up. This should only show up if 1) User is a
third party or has no professional relationship to the business (user
types 3 and 4) 2) Only if this information has been provided. At present
this comes up even though it wasn’t a question I encountered

Another example. If I do not know is selected for supplier, it shows as
blank fields. This should appear as ‘I do not know’ for location of
supplier. And ‘name and address of supplier’ shouldn’t appear

Website and Additional contact details haven’t been provided, but this
still shows up as a blank option. This doesn’t need to appear if there
is no data.
Last example where this was observed. No-documents uploaded, there
doesn’t need to be a field here.

2) Change link journeys, a couple of the change link journeys weren’t
quite right.

When you change the location of end users, it should take you to the
page titled ‘Where were the goods, services, technological assistance or
technology supplied to?’ However, this takes you to the page titled
Where were the goods, services, technological assistance or technology
supplied from?

---------

Co-authored-by: Morgan Rees <morganrees@DBT000491.local>